requirements.txt
django==6.0.4
python-slugify
